LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-base-appearance-computed-style-expected.txt

button
SIBLING

FAIL UA styles of base appearance <select>. assert_equals: min-inline-size expected "calc-size(auto, max(size, 24px))" but got "0px"
FAIL UA styles of base appearance select::picker-icon. assert_equals: content expected "counter(fake-counter-name, disclosure-open)" but got "\"▼\" / \"\""
FAIL UA styles of base appearance ::picker(select) assert_equals: box-sizing expected "border-box" but got ""
PASS UA styles of base appearance <option>.
PASS UA styles of base appearance option::checkmark.
FAIL UA styles of base appearance <optgroup>. assert_equals: display expected "block" but got "inline"
FAIL UA styles of base appearance <legend>. assert_equals: min-block-size expected "13px" but got "0px"
FAIL UA styles of base appearance select <button>. assert_equals: display expected "contents" but got "block"
PASS UA styles of base appearance <option> in <optgroup>.
